refactor: Replace GuestNavbar with Navbar component for consistent layout across pages
